fix: 後台座位管理遷移到 Pinia
